server/requirements.txt
# PrepOS Backend Dependencies
# Python 3.10+ required

# Web Framework - use latest FastAPI compatible with google-adk
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Database
pymongo>=4.6.0
motor>=3.3.0  # Async MongoDB driver

# Google AI / ADK - let it manage its own dependencies
google-generativeai>=0.8.0
google-adk>=1.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
authlib>=1.5.1
httpx>=0.27.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
